feat(academic-affairs):Update Grade student

+++ src/main/java/org/example/quanlytrungtam/grade/Grades.java
@@ -17,7 +17,7 @@ public class Grades {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "gradeid")
-    private Long gradeId;
+    private Integer gradeId;
     @ManyToOne
     @JoinColumn(name = "studentid", nullable = false)
     private Student student;
